Fix case where device goes to sleep with LOW
If the flywheel stops with the magnet triggering the sensor continuously
(i.e. setting it LOW) when the device goes to sleep wake up is
immediately triggered and goes into a continuos loop preventing actual
sleep. Checking the state and setting the makeup input level to the
opposite as the sensor reading can prevent this behaviour