Implement linear regression DF calculator
The linear regression DF calculation model seems less error prone to
inaccurate deltaTime values at the start and end of the phases. The
issue is that there are deltaTimes that get included in the Recovery
phase that should be still within the Drive. These errors are systematic
and due to a the reduction of the force applied to the flywheel by the
end of the Drive phase.

- Implement linear regression model
- Add LinearRegressorService to keep track of the running sums
- Add functions to the service necessary for DF calculator
- Update DF calculation filter conditions in StrokeService
- Clean up pervious start-end velocity model related variables and code
- Update tests