Change BLE broadcast method to interval
Decouple the BLE notification timing from the rotation events and use an
approx. one second interval to notify clients and use new stroke or 4sec
fixed time to update the data to be broadcasted on the next notify event
This should ensure smoother and better data quality (less spiky) on the
clients side.