Improve TS Quadratic algorithm
Improve the accuracy of the Theil-Sen Quadratic regression algorithm by
removing the intermediary median step when creating the "A" values. This
approach requires one less sorting but significantly higher vector size.
However, based on testing calculating the intermediary median is less
performant due to its higher time complexity compared to sorting a
bigger vector, so the ESP32 hardware is able to handle the increased
accuracy.

Fine tune existing profiles to accommodate the changes to the algorithm
as well as update tests.