Improve memory usage by using const reference
Pass the rowing data via const reference to save some memory by avoiding
the creation of multiple copy of this struct. Limit the amount of data
points that are collected during the recovery phase as well as reset the
collected data as soon as possible to avoid unnecessarily high heep
usage causing potential crash.