Rework WebSocket broadcasts to improve performance
To improve performance of the data broadcasts separates the settings and
rowing data (that needs update on a per stroke basis) to avoid sending
unnecessary data on every stroke. Broadcast settings on connect (as well
as on settings change) and on new battery level measurement.

Use RTOS xTask to offload the creation fo the data string as well as the
broadcast to the clients (testing showed that this could take several
hundreds of milliseconds). Use binary to broadcast the data as it is
slightly more performant.

This includes breaking changes to the WebSocket API.