Offload metric BLE notifies to second core
Move the notify of the metrics to the second core to improve performance
and reduce the chance of impulses not processed correctly with machines
having a high number of impulses per rotation.

Frees the main loop from most time consuming task other than running the
algorithm and calculating metrics (apart from - based on benchmarks the
negligible overhead of creating tasks).