Add interfaces to support dependency injection
Implement pure abstract classes as interfaces in order to enable mocking
via DI. This would allow extending test coverage that supports long term
maintenance.

Before settling with virtual pure abstract classes as interfaces testing
and extensive research have been conducted. It has been analyzed and
tested that the template based (pure compile time) approach does not
meet the necessary requirements due to the following reasons:

It requires concepts to enable intellisense which needs newer compiler
which is not yet officially supported by the used stack.

Also extensive benchmarking showed that the new compiler (v12) is less
performant than the current (v8). This was marginally with approx.
20-25us though, so while the interface approach added approx 16-20us
overhead it equaled to the overhead added by the new compiler.

Compile time templated approach is more complicated with the selected
mocking library as it still needs virtual methods on the class to be
able to stub/mock dependencies. So on the unit test level it requires
significant boilerplate code.