Improve failsafe for torque based stroke detection
Reset drag coefficient, end the drive and start the recovery phase
when a drive phase is too long (the driveHandleForces vector exceeds the
driveHandleForcesMaxCapacity) when stroke detection is torque based.

This is necessary in those machines that has a lot of bearing and other
friction causing significant difference in drag factor for lower and
higher rotation speeds and the recovery is not long enough for the
minimumRecoverySlopeMargin to kick in.