Experimental support for OldDanube Kayak Ergs
Adding experimental initial support for OldDanube Kayak Ergs with a 6-magnet
that are currently being developed and tested with the OldDanube team.

These settings are subject change in the future as the hardware is
developed more data is collected and analyzed.

Add new board profiles supporting the `Waveshare ESP32-S3 Zero` and
`FireBeetle 2 ESP32-S3` boards. Update documentation to reflect
these changes.