Relax dragCoefficient calculation requirements
This change relaxes the requirements for calculating the drag factor as
it allows for a smaller number of recovery delta times to be used
when the drag coefficient is zero. This improves the initial detection
when the drag coefficient is not yet established or when it has been
reset due to the drive phase exceeding the maximum drive handle forces
capacity.